remove PromptTemplateService
